[feat/#154] SolrouteMap 컴포넌트에 객체 정보 리스트 전달
